# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ModelHub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Provider {
  id: number;
  name: string;
  providerType: string;
  /** @nullable */
  apiKey?: string | null;
  /** @nullable */
  baseUrl: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  customHeaders?: string | null;
  /** @nullable */
  organizationId?: string | null;
  /** @nullable */
  endpoint?: string | null;
  enabled: boolean;
  /** @nullable */
  detectedType?: string | null;
  /** @nullable */
  confidence?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProviderInput {
  /** @minLength 1 */
  name: string;
  providerType: string;
  apiKey?: string;
  baseUrl?: string;
  model?: string;
  customHeaders?: string;
  organizationId?: string;
  endpoint?: string;
  enabled?: boolean;
}

export interface ProviderPatch {
  /** @minLength 1 */
  name?: string;
  providerType?: string;
  apiKey?: string;
  baseUrl?: string;
  model?: string;
  customHeaders?: string;
  organizationId?: string;
  endpoint?: string;
  enabled?: boolean;
}

export interface DetectionInput {
  apiKey?: string;
  baseUrl?: string;
  endpoint?: string;
}

export interface DetectionResult {
  detected: boolean;
  /** @nullable */
  providerType?: string | null;
  /** @nullable */
  name?: string | null;
  confidence: number;
  /** @nullable */
  suggestedBaseUrl?: string | null;
  /** @nullable */
  reason?: string | null;
}

export interface TestResult {
  success: boolean;
  latencyMs: number;
  /** @nullable */
  message?: string | null;
  modelsEndpointWorking?: boolean;
}

export interface DiscoveredModel {
  id: string;
  name: string;
  /** @nullable */
  contextLength?: number | null;
  supportsVision?: boolean;
  supportsReasoning?: boolean;
  supportsFunctionCalling?: boolean;
  supportsImageGeneration?: boolean;
  supportsAudio?: boolean;
  supportsEmbeddings?: boolean;
  supportsStreaming?: boolean;
  supportsJsonMode?: boolean;
}

export interface Model {
  id: number;
  providerId: number;
  modelId: string;
  name: string;
  /** @nullable */
  contextLength?: number | null;
  supportsVision?: boolean;
  supportsReasoning?: boolean;
  supportsFunctionCalling?: boolean;
  supportsImageGeneration?: boolean;
  supportsAudio?: boolean;
  supportsEmbeddings?: boolean;
  supportsStreaming?: boolean;
  supportsJsonMode?: boolean;
  createdAt: string;
}

export interface ModelInput {
  providerId: number;
  modelId: string;
  name: string;
  contextLength?: number;
  supportsVision?: boolean;
  supportsReasoning?: boolean;
  supportsFunctionCalling?: boolean;
  supportsImageGeneration?: boolean;
  supportsAudio?: boolean;
  supportsEmbeddings?: boolean;
  supportsStreaming?: boolean;
  supportsJsonMode?: boolean;
}

export interface ModelPatch {
  name?: string;
  contextLength?: number;
  supportsVision?: boolean;
  supportsReasoning?: boolean;
  supportsFunctionCalling?: boolean;
  supportsImageGeneration?: boolean;
  supportsAudio?: boolean;
  supportsEmbeddings?: boolean;
  supportsStreaming?: boolean;
  supportsJsonMode?: boolean;
}

export interface Message {
  role: string;
  content: string;
}

export interface ChatRequest {
  providerId: number;
  model: string;
  messages: Message[];
  systemPrompt?: string;
  temperature?: number;
  topP?: number;
  maxTokens?: number;
  presencePenalty?: number;
  frequencyPenalty?: number;
  /** @nullable */
  seed?: number | null;
  stop?: string[];
  /** @nullable */
  responseFormat?: string | null;
  /** @nullable */
  tools?: string | null;
}

export interface TokenUsage {
  promptTokens?: number;
  completionTokens?: number;
  totalTokens?: number;
}

export interface ChatResponse {
  content: string;
  latencyMs: number;
  usage?: TokenUsage;
  /** @nullable */
  costEstimate?: number | null;
  /** @nullable */
  rawResponse?: string | null;
  /** @nullable */
  rawRequest?: string | null;
}

export interface HistoryEntry {
  id: number;
  providerId: number;
  /** @nullable */
  providerName?: string | null;
  model: string;
  requestType: string;
  status: string;
  latencyMs: number;
  /** @nullable */
  promptTokens?: number | null;
  /** @nullable */
  completionTokens?: number | null;
  /** @nullable */
  totalTokens?: number | null;
  /** @nullable */
  costEstimate?: number | null;
  starred: boolean;
  /** @nullable */
  rawRequest?: string | null;
  /** @nullable */
  rawResponse?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
}

export interface HistoryPage {
  items: HistoryEntry[];
  total: number;
}

export interface HistoryInput {
  providerId: number;
  model: string;
  requestType: string;
  status: string;
  latencyMs: number;
  promptTokens?: number;
  completionTokens?: number;
  totalTokens?: number;
  costEstimate?: number;
  rawRequest?: string;
  rawResponse?: string;
  errorMessage?: string;
}

export interface HistoryPatch {
  starred?: boolean;
  errorMessage?: string;
}

export interface Settings {
  id: number;
  /** @nullable */
  defaultProviderId?: number | null;
  /** @nullable */
  defaultModel?: string | null;
  streamingEnabled: boolean;
  autoSave: boolean;
  retryCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface SettingsInput {
  defaultProviderId?: number;
  defaultModel?: string;
  streamingEnabled?: boolean;
  autoSave?: boolean;
  retryCount?: number;
}

export interface ProviderStat {
  providerName: string;
  requestCount: number;
  tokensUsed?: number;
  costEstimate?: number;
}

export interface DashboardStats {
  connectedProviders: number;
  availableModels: number;
  requestsToday: number;
  tokensToday: number;
  costToday: number;
  requestsByProvider?: ProviderStat[];
  recentActivity?: HistoryEntry[];
}

export type ListModelsParams = {
providerId?: number;
};

export type ListHistoryParams = {
search?: string;
providerId?: number;
starred?: boolean;
limit?: number;
offset?: number;
};