# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ModelHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /models:
    get:
      operationId: listModels
      tags: [models]
      summary: List all saved models
      parameters:
        - name: providerId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Model"
    post:
      operationId: createModel
      tags: [models]
      summary: Save a model configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelInput"
      responses:
        "201":
          description: Created model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
  /models/{id}:
    patch:
      operationId: updateModel
      tags: [models]
      summary: Update a saved model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelPatch"
      responses:
        "200":
          description: Updated model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
    delete:
      operationId: deleteModel
      tags: [models]
      summary: Delete a saved model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Playground ---
  /playground/chat:
    post:
      operationId: playgroundChat
      tags: [playground]
      summary: Send a chat completion request through a provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /playground/stream:
    post:
      operationId: playgroundStream
      tags: [playground]
      summary: Stream a chat completion (SSE)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: string

  # --- History ---
  /history:
    get:
      operationId: listHistory
      tags: [history]
      summary: List request history
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: providerId
          in: query
          required: false
          schema:
            type: integer
        - name: starred
          in: query
          required: false
          schema:
            type: boolean
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: History list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryPage"
    post:
      operationId: createHistory
      tags: [history]
      summary: Save a new history entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HistoryInput"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryEntry"
  /history/{id}:
    get:
      operationId: getHistory
      tags: [history]
      summary: Get a single history entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: History entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateHistory
      tags: [history]
      summary: Update a history entry (star/unstar, notes)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HistoryPatch"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryEntry"
    delete:
      operationId: deleteHistory
      tags: [history]
      summary: Delete a history entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Settings ---
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # --- Provider ---
...
    Model:
      type: object
      required: [id, providerId, modelId, name, createdAt]
      properties:
        id:
          type: integer
        providerId:
          type: integer
        modelId:
          type: string
        name:
          type: string
        contextLength:
          type: ["integer", "null"]
        supportsVision:
          type: boolean
        supportsReasoning:
          type: boolean
        supportsFunctionCalling:
          type: boolean
        supportsImageGeneration:
          type: boolean
        supportsAudio:
          type: boolean
        supportsEmbeddings:
          type: boolean
        supportsStreaming:
          type: boolean
        supportsJsonMode:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ModelInput:
      type: object
      required: [providerId, modelId, name]
      properties:
        providerId:
          type: integer
        modelId:
          type: string
        name:
          type: string
        contextLength:
          type: integer
        supportsVision:
          type: boolean
        supportsReasoning:
          type: boolean
        supportsFunctionCalling:
          type: boolean
        supportsImageGeneration:
          type: boolean
        supportsAudio:
          type: boolean
        supportsEmbeddings:
          type: boolean
        supportsStreaming:
          type: boolean
        supportsJsonMode:
          type: boolean
    ModelPatch:
      type: object
      properties:
        name:
          type: string
        contextLength:
          type: integer
        supportsVision:
          type: boolean
        supportsReasoning:
          type: boolean
        supportsFunctionCalling:
          type: boolean
        supportsImageGeneration:
          type: boolean
        supportsAudio:
          type: boolean
        supportsEmbeddings:
          type: boolean
        supportsStreaming:
          type: boolean
        supportsJsonMode:
          type: boolean

    # --- Playground ---
    Message:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
        content:
          type: string
    ChatRequest:
      type: object
      required: [providerId, model, messages]
      properties:
        providerId:
          type: integer
        model:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        systemPrompt:
          type: string
        temperature:
          type: number
        topP:
          type: number
        maxTokens:
          type: integer
        presencePenalty:
          type: number
        frequencyPenalty:
          type: number
        seed:
          type: ["integer", "null"]
        stop:
          type: array
          items:
            type: string
        responseFormat:
          type: ["string", "null"]
        tools:
          type: ["string", "null"]
    TokenUsage:
      type: object
      properties:
        promptTokens:
          type: integer
        completionTokens:
          type: integer
        totalTokens:
          type: integer
    ChatResponse:
      type: object
      required: [content, latencyMs]
      properties:
        content:
          type: string
        latencyMs:
          type: integer
        usage:
          $ref: "#/components/schemas/TokenUsage"
        costEstimate:
          type: ["number", "null"]
        rawResponse:
          type: ["string", "null"]
        rawRequest:
          type: ["string", "null"]

    # --- History ---
    HistoryEntry:
      type: object
      required: [id, providerId, model, requestType, status, latencyMs, starred, createdAt]
      properties:
        id:
          type: integer
        providerId:
          type: integer
        providerName:
          type: ["string", "null"]
        model:
          type: string
        requestType:
          type: string
        status:
          type: string
        latencyMs:
          type: integer
        promptTokens:
          type: ["integer", "null"]
        completionTokens:
          type: ["integer", "null"]
        totalTokens:
          type: ["integer", "null"]
        costEstimate:
          type: ["number", "null"]
        starred:
          type: boolean
        rawRequest:
          type: ["string", "null"]
        rawResponse:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    HistoryPage:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
        total:
          type: integer
    HistoryInput:
      type: object
      required: [providerId, model, requestType, status, latencyMs]
      properties:
        providerId:
          type: integer
        model:
          type: string
        requestType:
          type: string
        status:
          type: string
        latencyMs:
          type: integer
        promptTokens:
          type: integer
        completionTokens:
          type: integer
        totalTokens:
          type: integer
        costEstimate:
          type: number
        rawRequest:
          type: string
        rawResponse:
          type: string
        errorMessage:
          type: string
    HistoryPatch:
      type: object
      properties:
        starred:
          type: boolean
        errorMessage:
          type: string

    # --- Settings ---
...