# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ModelHub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from './message';

export interface ChatRequest {
  providerId: number;
  model: string;
  messages: Message[];
  systemPrompt?: string;
  temperature?: number;
  topP?: number;
  maxTokens?: number;
  presencePenalty?: number;
  frequencyPenalty?: number;
  /** @nullable */
  seed?: number | null;
  stop?: string[];
  /** @nullable */
  responseFormat?: string | null;
  /** @nullable */
  tools?: string | null;
}