# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ModelHub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TokenUsage } from './tokenUsage';

export interface ChatResponse {
  content: string;
  latencyMs: number;
  usage?: TokenUsage;
  /** @nullable */
  costEstimate?: number | null;
  /** @nullable */
  rawResponse?: string | null;
  /** @nullable */
  rawRequest?: string | null;
}