# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ModelHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryEntry {
  id: number;
  providerId: number;
  /** @nullable */
  providerName?: string | null;
  model: string;
  requestType: string;
  status: string;
  latencyMs: number;
  /** @nullable */
  promptTokens?: number | null;
  /** @nullable */
  completionTokens?: number | null;
  /** @nullable */
  totalTokens?: number | null;
  /** @nullable */
  costEstimate?: number | null;
  starred: boolean;
  /** @nullable */
  rawRequest?: string | null;
  /** @nullable */
  rawResponse?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: Date;
}